adds serializable pipeline attribute to Dataset
